chore: add assigned user to prisma issue schema
Create a relation between issues and users. An issue can be assigned to
one user need to create a relation relating the key in issues to the id
referenced in users. Also need to add another key to users model that
points to the list of issues that has been assigned to that user.